move extract() to conftest
