+++ test/conftest.py
@@ -3,11 +3,14 @@ from utility.helpers import conn_pool
 from src.config import TestingConfig
 from flask import g
 import pytest
+import json
+
 
 @pytest.fixture(autouse=True, scope='module')
 def _app_runner():
     app = creat_app(TestingConfig)
     yield app
+
 
 @pytest.fixture(autouse=True, scope='module')
 def client(_app_runner):
@@ -21,3 +24,9 @@ def client(_app_runner):
             conn.rollback()
             conns.putconn(conn)
 
+
+def extract(response):
+    info = json.loads(response.data)['info']
+    data = json.loads(response.data)['data']
+    error = json.loads(response.data)['error']
+    return info, data, error
